调整迁移表结构

+++ migrations/2018_01_16_163058_create_sql_records_table.php
@@ -16,7 +16,7 @@ class CreateSqlRecordsTable extends Migration
         Schema::connection('stat_mysql')->create('sql_records', function (Blueprint $table) {
             $table->increments('id');
             $table->text('sql')->comment('sql记录');
-            $table->unsignedInteger('time')->comment('sql执行时间(秒)');
+            $table->float('time',3,3)->comment('sql执行时间(秒)');
             $table->timestamps();
         });
     }
